# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_financeos_api_internal_domain_entity.Account:
    properties:
      balance:
        type: number
      color:
        type: string
      created_at:
        type: string
      credit_limit:
        type: number
      icon:
        type: string
      id:
        type: string
      institution:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        description: checking, savings, credit_card, investment, wallet, other
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_financeos_api_internal_usecase.CreateAccountRequest:
    properties:
      balance:
        type: number
      color:
        type: string
      credit_limit:
        type: number
      icon:
        type: string
      institution:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      type:
        enum:
        - checking
        - savings
        - credit_card
        - investment
        - wallet
        - other
        type: string
    required:
    - name
    - type
    type: object
...
  github_com_financeos_api_internal_usecase.UpdateAccountRequest:
    properties:
      color:
        type: string
      credit_limit:
        type: number
      icon:
        type: string
      institution:
        type: string
      name:
        type: string
    type: object
...
  internal_handler.AIChatData:
    properties:
      response:
        example: Você gastou R$1.500 em Alimentação este mês, 25% acima da média.
        type: string
    type: object
  internal_handler.AIChatResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.AIChatData'
    type: object
  internal_handler.AIForecastResponse:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
  internal_handler.AccountListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Account'
        type: array
    type: object
  internal_handler.AccountResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Account'
    type: object
  internal_handler.AccountSummaryData:
    properties:
      account_balances:
        additionalProperties:
          format: float64
          type: number
        type: object
      total_balance:
        example: 15000
        type: number
    type: object
  internal_handler.AccountSummaryResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.AccountSummaryData'
    type: object
...
  internal_handler.ErrorDetail:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: name is required
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
    type: object
...
  internal_handler.MessageData:
    properties:
      message:
        example: logged out successfully
        type: string
    type: object
  internal_handler.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.MessageData'
    type: object
...
host: localhost:8000
info:
  contact:
    name: FinanceOS Support
    url: https://github.com/financeos
  description: API completa para gestão financeira pessoal — contas, transações, orçamentos,
    investimentos, metas e IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinanceOS API
  version: 1.0.0
paths:
  /accounts:
    get:
      description: Retorna todas as contas ativas do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar contas
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Cria uma nova conta bancária, carteira ou cartão
      parameters:
      - description: Dados da conta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar conta
      tags:
      - Accounts
  /accounts/{id}:
    delete:
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Desativar conta
      tags:
      - Accounts
    get:
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar conta por ID
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar conta
      tags:
      - Accounts
  /accounts/summary:
    get:
      description: Retorna saldo total e saldo por conta
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resumo de saldo
      tags:
      - Accounts
  /ai/chat:
    post:
      consumes:
      - application/json
      description: Conversa com o assistente financeiro alimentado por Claude
      parameters:
      - description: Mensagem (máx. 2000 chars)
        in: body
        name: body
        required: true
        schema:
          properties:
            message:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assistente financeiro
      tags:
      - AI
  /ai/portfolio-analysis:
    get:
      description: Insights sobre diversificação e performance (plano Pro)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIForecastResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Análise de portfólio com IA
      tags:
      - AI
  /ai/spending-forecast:
    get:
      description: Análise preditiva de gastos com IA (plano Pro)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIForecastResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Previsão de gastos
      tags:
      - AI
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"